Update Read.me
The Read me is being updated inline with the Corretto plan for the coming 7 weeks. By the end of this current road map, all the arithmetic curve operations will be completed. Following this, an assessment of how Corretto will be furthered, as per the purpose of the library as for other projects, will be given.