Implement Kalinski Modular inverse algorithm
**Refecences:**
B. S. Kaliski Jr. - The  Montgomery  inverse  and  its  applica-tions.
IEEE Transactions on Computers, 44(8):1064–1065, August-1995

After this commit, #9 can be closed.

-Implemented Kalinski Modular Inverse
algorithm.
- Implemented tests for the algorithm
which are passing successfully.
- Improved the docs of the inverse function.

At this point, Implement the Savas & Koç
modification of this algorithm to increase
the performance of the operation will be great.
So maybe we can implement the algorithm and then
benchmark coth solutions to see which performs
better with our `FieldElement` definition.

Tools for the second algorithm implementation
were builded on 54152f2 and allowed us to close #11