Refactor point doubling algorithm
Until we don't find which is the reason that
makes the algorithm fail (since it's implemented
as the pseudocode says on the `Twisted Edwards
Curves revisted` paper) point doubling will be
implemented as an addition of a point to itself.

This is also correct, but forces us to perform 1M
and 1S extra. So we will work to solve it.