Impl `new_from_y_coord()` for `ProjectivePoint`
This closes #57 and it's what we needed to close #51.

This function tries to build a Point over the Doppio Curve from
a `Y` coordinate and a Choice that determines the Sign o the `X`
coordinate that the user wants to use.

The function gets `X` by solving:
`+-X = mod_sqrt((y^2 -1)/(dy^2 - a))`.

The sign of `x` is choosen with a `Choice` parameter.
```
 Choice(0) ->
 Choice(1) ->
```
Then Z is always equal to `1`.

 Returns
`Some(ProjectivePoint)` if there exists a result for the `mod_sqrt`
and `None` if the resulting `x^2` isn't a QR modulo `FIELD_L`.