Impl `Eq` `PartialEq` and `ConstantTimeEq` for `EdwardsPoint`
`Eq` just extends `PartialEq` which calls
the `ConstantTimeEq` implementation.

The last one, compares the two coordinates of the
`EdwardsPoint` in `Affine Coordinates` in constant
time and performs an AND operation between the
resulting Choices.

The result is returned as a Choice where 0 = false
and 1 = true.

So the Extended Edwards equalty testing relies on the
`AffinePoint` equalty implementation.