Impl `Eq` `PartialEq` &  `ConstantTimeEq` for `ProjectivePoint`
`Eq` just extends `PartialEq` which calls
the `ConstantTimeEq` implementation.

The last one, compares the two coordinates of
both compared PorjectivePoints moved as
a AffinePoint in constant time and performs an
AND operation between the resulting Choices.

The result is returned as a Choice where 0 = false
and 1 = true.

Also clean `Doubling` tests since we can now guarantee
that it's passing. Which almos finnishes with #50