Base point computation
This commit is coupled to both a verification
on the basepoint computation as well as a
proof that the computed base point doesn't
contain a modular square root. Additionally,
it is stated which output of a modular square
root operation is taken to be positive. Which
is crucial for sign choice operations. However,
this is only a conjecture and a formal proof
will need to be created.