Implement new Point Ops skeleton & `Doubling` for `RistrettoPoint`
Since some point operations need the map between the choosen
isomorphic twist and Doppio (and we don't have computed it yet).

- Doubling is the only operation that has been both
implemented and tested.

- Mul is implemented with the `double_and_add` algorithm.
It relies on addition and doubling, so due to the first
dependency, it cannot be tested yet.

- Addition has not been implemented yet since we need to figure
out the mapping between the isomorphic twist addition and the
Doppio addition (both in TwEds Extended Coords)