Compress `EdwardsPoint` from its Affine coords
Refactor the compression input variables from
Extended coordinates to Affine coordinates.

Compress in extended if Z != 1, causes
a possible bad encoding as seen with
the point:

(0, X, X, 0) with X != 1. This should be encoded
as the identity point but wasn't.

Now carrying the point to Affine coordinates first,
we avoid this problem and fix the Compression bug.