Implement tests for Random point generation validity.
- Refactored an error found on the
`elligator_ristretto_flavour()` function which
returned a `CompletePôint` so we needed to turn it
into an `EdwardsPoint` to encapsulate it onto a
`RistrettoPoint`.

- Added tests with `rand` feature to generate random
points by multiplying the basepoint by a random Scalar.

This has been done like this instead of generating
random `RistrettoPoints` directly since it caused the
tests to fail, meaning that something was wrong on the
random point generation using the Elligator hash-to-group
function.